KAFKA-9726: Make replication policy config not predefined
Before `MirrorClientConfig` and `MirrorConnectorConfig` had the `DefaultReplicationPolicy`'s configuration as a part of their own and `MirrorMakerConfig` used this to pass the configuration further. However, this is not well aligned with the pluggably nature of `ReplicationPolicy` classes as the set of configurations cannot be known in advance.

This commit removes explicit definition of `replication.policy.separator` from `MirrorClientConfig` and `MirrorConnectorConfig` configurations and makes `MirrorMakerConfig` pass replication policy configuration based on prefix.